Use version reported in Client Hello for fingerprint
The `SSL_version` function from OpenSSL returns 0x304 (772) for TLS
connections` using TLS v1.3.

This value differs from the one used by the client in the Client Hello
package, which is always 0x303 (771), even if the exchange is using TLS
v.13.

We really want to capture the value sent by the client for
fingerprinting, since we don't want to have varying fingerprints based
on the TLS version finally used in the exchange. Hence, we use the
`SSL_client_version` function to capture that information.